build/docker: remove RPM-specific code from DockerRun
Since we were only building RPM packages, we've made the questionable
decision to move some repeated RPM-specific code directly into the
DockerRun class.
That will no longer works as we will now reuse this class to run a
Debian builder container, so let's move this code outside.

Refs: #1472
